Add NPE data generator and training notebook.
At this point the NPE works but has the mysterious effect of floating to one corner of the screen. Not sure what's happening here but we suspect, based on the distrubution of the training data, that Box2D is messing up. It might make sense to not use Box2D and shift to Chipmunk, which has better models for fully elastic collisions.